feat: Introduce BLCS generator configuration helpers and refactor dataset generation
