+++ src/tasks/blcs/scripts/generate_dataset.py
@@ -28,18 +28,12 @@ from hydra.utils import to_absolute_path
 from omegaconf import DictConfig, OmegaConf
 from tqdm.auto import tqdm
 
+from src.tasks.blcs.generate_dataset.config import build_generator_config
 from src.tasks.blcs.generate_dataset.io.dataset_io import BLCSDatasetWriter
 from src.tasks.blcs.generate_dataset.scene_generator import (
     GeneratorConfig,
 )
-from src.tasks.blcs.generate_dataset.simulation.ball_physics import PhysicsConfig
-from src.tasks.blcs.generate_dataset.simulation.rally_simulator import RallyConfig
-from src.tasks.blcs.generate_dataset.simulation.targeted_velocity_sampler import (
-    TargetedVelocityConfig,
-)
 from src.tasks.blcs.generate_dataset.utils.parallel_runner import generate_parallel_scenes
-from src.utils.projection.camera_projector import CameraConfig
-from src.utils.schema.court import CourtConfig
 
 logging.basicConfig(
     level=logging.INFO,
@@ -54,107 +48,6 @@ def _seed_everything(seed: int) -> None:
     random.seed(seed)
     np.random.seed(seed)
     torch.manual_seed(seed)
-
-
-def build_generator_config(cfg: DictConfig) -> GeneratorConfig:
-    physics_config = PhysicsConfig(
-        gravity=float(cfg.physics.gravity),
-        k_drag=float(cfg.physics.k_drag),
-        k_magnus=float(cfg.physics.k_magnus),
-        e_z=float(cfg.physics.e_z),
-        mu=float(cfg.physics.mu),
-        alpha_net=float(cfg.physics.alpha_net),
-        alpha_net_cord=float(cfg.physics.alpha_net_cord),
-        alpha_fence=float(cfg.physics.alpha_fence),
-        net_half_thickness=float(cfg.physics.net_half_thickness),
-        net_cord_radius=float(cfg.physics.net_cord_radius),
-        dt=float(cfg.physics.dt),
-        use_drag=bool(cfg.physics.use_drag),
-        use_magnus=bool(cfg.physics.use_magnus),
-        wind=tuple(cfg.physics.wind),
-        gravity_range=tuple(cfg.physics.gravity_range),
-        k_drag_range=tuple(cfg.physics.k_drag_range),
-        k_magnus_range=tuple(cfg.physics.k_magnus_range),
-        e_z_range=tuple(cfg.physics.e_z_range),
-        mu_range=tuple(cfg.physics.mu_range),
-        wind_speed_range=tuple(cfg.physics.wind_speed_range),
-        wind_direction_range_deg=tuple(cfg.physics.wind_direction_range_deg),
-    )
-
-    rally_config = RallyConfig(
-        z_range=tuple(cfg.rally.z_range),
-        spin_x_range=tuple(cfg.rally.spin_x_range),
-        spin_y_range=tuple(cfg.rally.spin_y_range),
-        spin_z_range=tuple(cfg.rally.spin_z_range),
-        max_sim_frames=int(cfg.rally.max_sim_frames),
-        output_fps=int(cfg.rally.output_fps),
-        sim_fps=int(cfg.rally.sim_fps),
-        max_rallies=int(cfg.rally.max_rallies),
-        max_total_frames=int(cfg.rally.max_total_frames),
-        hit_timing_range=tuple(cfg.rally.hit_timing_range),
-        return_z_range=tuple(cfg.rally.return_z_range),
-        serve_probability=float(cfg.rally.serve_probability),
-        serve_z_range=tuple(cfg.rally.serve_z_range),
-        toss_vz_range=tuple(cfg.rally.toss_vz_range),
-        toss_xy_noise_range=tuple(cfg.rally.toss_xy_noise_range),
-        toss_max_frames=int(cfg.rally.toss_max_frames),
-        toss_z0_tolerance=float(cfg.rally.toss_z0_tolerance),
-        volley_probability=float(cfg.rally.volley_probability),
-        normal_return_probability=float(cfg.rally.normal_return_probability),
-        late_return_probability=float(cfg.rally.late_return_probability),
-    )
-
-    camera_config = CameraConfig(
-        z_min=float(cfg.camera.z_min),
-        z_max=float(cfg.camera.z_max),
-        hfov_deg=float(cfg.camera.hfov_deg),
-        image_size=tuple(cfg.camera.image_size),
-        fixed_look_at=tuple(cfg.camera.fixed_look_at),
-        fixed_baseline_clear_extra=float(cfg.camera.fixed_baseline_clear_extra),
-        fixed_position_noise_radius=float(cfg.camera.fixed_position_noise_radius),
-        fixed_look_at_xy_radius=float(cfg.camera.fixed_look_at_xy_radius),
-    )
-
-    # Court config
-    court_cfg = cfg.generator.court
-    court_config = CourtConfig(
-        net_post_offset_x=float(court_cfg.net_post_offset_x),
-        net_post_offset_x_range=tuple(court_cfg.net_post_offset_x_range),
-    )
-
-    targeted_velocity_config = TargetedVelocityConfig(
-        azimuth_noise_deg=float(cfg.targeted_velocity.azimuth_noise_deg),
-        elevation_noise_deg=float(cfg.targeted_velocity.elevation_noise_deg),
-        speed_variation=float(cfg.targeted_velocity.speed_variation),
-        min_elevation_deg=float(cfg.targeted_velocity.min_elevation_deg),
-        max_elevation_deg=float(cfg.targeted_velocity.max_elevation_deg),
-        drive_elevation_range_deg=tuple(cfg.targeted_velocity.drive_elevation_range_deg),
-        lob_elevation_range_deg=tuple(cfg.targeted_velocity.lob_elevation_range_deg),
-        lob_probability=float(cfg.targeted_velocity.lob_probability),
-        min_speed=float(cfg.targeted_velocity.min_speed),
-        max_speed=float(cfg.targeted_velocity.max_speed),
-        gravity=float(cfg.targeted_velocity.gravity),
-        speed_solve_max_iters=int(cfg.targeted_velocity.speed_solve_max_iters),
-        speed_solve_tol=float(cfg.targeted_velocity.speed_solve_tol),
-        refine_enabled=bool(cfg.targeted_velocity.refine_enabled),
-        refine_iters=int(cfg.targeted_velocity.refine_iters),
-        refine_speed_scale_min=float(cfg.targeted_velocity.refine_speed_scale_min),
-        refine_speed_scale_max=float(cfg.targeted_velocity.refine_speed_scale_max),
-        refine_max_azimuth_adjust_deg=float(cfg.targeted_velocity.refine_max_azimuth_adjust_deg),
-        refine_max_frames=int(cfg.targeted_velocity.refine_max_frames),
-        net_clearance_enabled=bool(cfg.targeted_velocity.net_clearance_enabled),
-        net_clearance_min=float(cfg.targeted_velocity.net_clearance_min),
-        net_clearance_max_attempts=int(cfg.targeted_velocity.net_clearance_max_attempts),
-        net_clearance_max_frames=int(cfg.targeted_velocity.net_clearance_max_frames),
-    )
-
-    return GeneratorConfig(
-        physics=physics_config,
-        rally=rally_config,
-        camera=camera_config,
-        targeted_velocity=targeted_velocity_config,
-        court=court_config,
-    )
 
 
 def _hydra_main(func: F) -> F:
